chore/matter-switch lazy loading can_handle refactor
